requirements.txt
PyQt6==6.7.1
pytesseract==0.3.13
pdf2image==1.17.0
pypdf==5.0.0
reportlab==4.2.2
Pillow==10.4.0
python-dotenv==1.0.1